cleanup ListController mixin
and move it into the controllers using it
