+++ ui/app/controllers/vault/cluster/access/leases/list.js
@@ -7,9 +7,10 @@ import { inject as service } from '@ember/service';
 import { computed } from '@ember/object';
 import Controller, { inject as controller } from '@ember/controller';
 import utils from 'vault/lib/key-utils';
-import ListController from 'core/mixins/list-controller';
+import escapeStringRegexp from 'escape-string-regexp';
+import commonPrefix from 'core/utils/common-prefix';
 
-export default Controller.extend(ListController, {
+export default Controller.extend({
   flashMessages: service(),
   store: service(),
   clusterController: controller('vault.cluster'),
@@ -23,10 +24,44 @@ export default Controller.extend(ListController, {
     };
   }),
 
+  queryParams: {
+    page: 'page',
+    pageFilter: 'pageFilter',
+  },
+
+  page: 1,
+  pageFilter: null,
+  filter: null,
+  filterFocused: false,
+
   isLoading: false,
 
   filterIsFolder: computed('filter', function () {
     return !!utils.keyIsFolder(this.filter);
+  }),
+
+  filterMatchesKey: computed('filter', 'model', 'model.[]', function () {
+    const { filter, model: content } = this;
+    return !!(content.length && content.findBy('id', filter));
+  }),
+
+  firstPartialMatch: computed('filter', 'model', 'model.[]', 'filterMatchesKey', function () {
+    const { filter, filterMatchesKey, model: content } = this;
+    const re = new RegExp('^' + escapeStringRegexp(filter));
+    const matchSet = content.filter((key) => re.test(key.id));
+    const match = matchSet[0];
+
+    if (filterMatchesKey || !match) {
+      return null;
+    }
+
+    const sharedPrefix = commonPrefix(content);
+    // if we already are filtering the prefix, then next we want
+    // the exact match
+    if (filter === sharedPrefix || matchSet.length === 1) {
+      return match;
+    }
+    return { id: sharedPrefix };
   }),
 
   emptyTitle: computed('baseKey.id', 'filter', 'filterIsFolder', function () {
@@ -46,6 +81,19 @@ export default Controller.extend(ListController, {
   }),
 
   actions: {
+    setFilter(val) {
+      this.set('filter', val);
+    },
+
+    setFilterFocus(bool) {
+      this.set('filterFocused', bool);
+    },
+
+    refresh() {
+      // bubble to the list-route
+      this.send('reload');
+    },
+
     revokePrefix(prefix, isForce) {
       const adapter = this.store.adapterFor('lease');
       const method = isForce ? 'forceRevokePrefix' : 'revokePrefix';
