+++ ui/app/controllers/vault/cluster/secrets/backend/list.js
@@ -10,9 +10,10 @@ import Controller from '@ember/controller';
 import utils from 'vault/lib/key-utils';
 import BackendCrumbMixin from 'vault/mixins/backend-crumb';
 import { task } from 'ember-concurrency';
-import ListController from 'core/mixins/list-controller';
+import escapeStringRegexp from 'escape-string-regexp';
+import commonPrefix from 'core/utils/common-prefix';
 
-export default Controller.extend(ListController, BackendCrumbMixin, {
+export default Controller.extend(BackendCrumbMixin, {
   navToNearestAncestor: task(function* (key) {
     const ancestors = utils.ancestorKeysForKey(key);
     let errored = false;
@@ -41,6 +42,36 @@ export default Controller.extend(ListController, BackendCrumbMixin, {
 
   flashMessages: service(),
   queryParams: ['page', 'pageFilter', 'tab'],
+  page: 1,
+  pageFilter: null,
+  filter: null,
+  filterFocused: false,
+
+  isLoading: false,
+
+  filterMatchesKey: computed('filter', 'model', 'model.[]', function () {
+    const { filter, model: content } = this;
+    return !!(content.length && content.findBy('id', filter));
+  }),
+
+  firstPartialMatch: computed('filter', 'model', 'model.[]', 'filterMatchesKey', function () {
+    const { filter, filterMatchesKey, model: content } = this;
+    const re = new RegExp('^' + escapeStringRegexp(filter));
+    const matchSet = content.filter((key) => re.test(key.id));
+    const match = matchSet[0];
+
+    if (filterMatchesKey || !match) {
+      return null;
+    }
+
+    const sharedPrefix = commonPrefix(content);
+    // if we already are filtering the prefix, then next we want
+    // the exact match
+    if (filter === sharedPrefix || matchSet.length === 1) {
+      return match;
+    }
+    return { id: sharedPrefix };
+  }),
 
   tab: '',
 
@@ -51,6 +82,19 @@ export default Controller.extend(ListController, BackendCrumbMixin, {
   isConfigurableTab: or('isCertTab', 'isConfigure'),
 
   actions: {
+    setFilter(val) {
+      this.set('filter', val);
+    },
+
+    setFilterFocus(bool) {
+      this.set('filterFocused', bool);
+    },
+
+    refresh() {
+      // bubble to the list-route
+      this.send('reload');
+    },
+
     chooseAction(action) {
       this.set('selectedAction', action);
     },
